Replace R2 with R3 renderer

+++ xray/Layers/xrRenderPC_R2/light_vis.cpp
@@ -36,10 +36,15 @@ void	light::vis_prepare			()
 	if (ps_r2_ls_flags.test(R2FLAG_EXP_DONT_TEST_UNSHADOWED) && !flags.bShadow)	skiptest=true;
 	if (ps_r2_ls_flags.test(R2FLAG_EXP_DONT_TEST_SHADOWED) && flags.bShadow)	skiptest=true;
 
+	//	TODO: DX10: Remove this pessimization
+	//skiptest	= true;
+
 	if (skiptest || Device.vCameraPosition.distance_to(spatial.sphere.P)<=(spatial.sphere.R*1.01f+safe_area))	{	// small error
 		vis.visible		=	true;
 		vis.pending		=	false;
 		vis.frame2test	=	frame	+ ::Random.randI(delay_small_min,delay_small_max);
+		//	TODO: DX10: Remove this pessimisation
+		//vis.frame2test	=	frame	+ 1;
 		return;
 	}
 
@@ -70,13 +75,18 @@ void	light::vis_update			()
 	if (!vis.pending)	return;
 
 	u32	frame			= Device.dwFrame;
-	u32 fragments		= RImplementation.occq_get	(vis.query_id);
+	u64 fragments		= RImplementation.occq_get	(vis.query_id);
 	//Log					("",fragments);
 	vis.visible			= (fragments > cullfragments);
 	vis.pending			= false;
-	if (vis.visible)	{
+	if (vis.visible)	
+	{
 		vis.frame2test	=	frame	+ ::Random.randI(delay_large_min,delay_large_max);
-	} else {
+		//	TODO: DX10: Remove this pessimisation
+		//vis.frame2test	=	frame	+ 1;
+	} 
+	else 
+	{
 		vis.frame2test	=	frame	+ 1; 
 	}
 }
